+++ modular/tensor_cloud.py
@@ -48,7 +48,7 @@ class TensorCloud:
 def input_tensor(size, name):
     if isinstance(size, int):
         size = [size]
-    return tf.placeholder(np.float64, [None] + size, name=name)
+    return tf.placeholder(np.float32, [None] + size, name=name)
 
 def flatten(tup):
     result = ()
